Issue #4586: Moved all resources for SummaryJavadocCheckTest
