+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/summaryjavadoc/InputSummaryJavadocCorrect.java
@@ -1,9 +1,9 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc;
+package com.puppycrawl.tools.checkstyle.checks.javadoc.summaryjavadoc;
 
 /**
  * Some Javadoc A {@code Foo} is a simple Javadoc.
  */
-class InputCorrectSummaryJavaDoc {
+class InputSummaryJavadocCorrect {
 
     /**
      * Some Javadoc This method returns.
